refacto: commands were cluttered +  no cli help
Solution: put all commands in a sub folder and use typer subcommands
Use typer.Option and typer.Argument to document each command parameter
